requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
pymupdf